ml-service/requirements.txt
Flask==3.0.0
easyocr==1.7.1
torch==2.1.1
torchvision==0.16.1
Pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.3.2
pandas==2.1.3
python-doctr[torch]
